Remove disqus, add anchors
Removing disqus, as there is a lot of tracking cookies added
as a result, and it's also the most expensive component to load.

I'm going to try handling comments some other way, such as through
Hackernews.

Adding a script to easily dev and test my blog posts stored in Dropbox.

Adding anchors to posts, to allow easy commenting and referencing.