Adding jekyll-postfiles plugins
It's a lot more intuitive to me to put files alongside the
blog post. So using a plugin that lets me do that.

There's also a caveat that, with the choice for semantic paths
and everything at the top level directory, the number of images
will get quite confusing. So putting a trailing slash in my blog
posts to better clarify which assets belong with which posts during
render.

note: with this plugin, images that have dates in them will not work:
https://github.com/nhoizey/jekyll-postfiles/issues/1#issuecomment-299209146